backend/requirements.txt
fastapi
uvicorn
sse-starlette
pydantic
python-dotenv
httpx
asyncio
langchain-groq
langchain
langchain-core
